requirements.txt
fastapi~=0.89.1
SQLAlchemy~=1.4.46
pydantic~=1.10.4
uvicorn~=0.20.0
python-dotenv~=0.21.1
psycopg2-binary~=2.9.5
alembic~=1.9.2
pytest~=7.2.1
httpx~=0.23.3
sqlalchemy_utils~=0.39.0
redis~=4.4.2